+++ response_operations_ui/templates/search-respondent.html
@@ -24,7 +24,7 @@
 
                 </div>
                 {% if response %}
-                    No Respondent found.
+                    {{ response }}
                 {% endif %}
             </div>
         </div>
